# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenFireblocks API
  description: |
    Enterprise-grade institutional cryptocurrency key management and signing platform.
    
    ## Authentication
    All requests require an API key in the `X-API-Key` header.
    
    ## Rate Limiting
    - Standard: 100 requests/second
    - Premium: 1000 requests/second
  version: 1.0.0
  contact:
    name: OpenFireblocks Support
    email: support@openfireblocks.io
servers:
  - url: https://api.openfireblocks.io/v1
    description: Production
  - url: https://api-staging.openfireblocks.io/v1
    description: Staging
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - apiKey: []
paths:
...
  /keys:
    get:
      summary: List key pairs
      tags: [Keys]
      responses:
        '200':
          description: List of key pairs
    post:
      summary: Create key pair via DKG
      tags: [Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: DKG ceremony initiated
  /keys/{key_id}:
    get:
      summary: Get key pair
      tags: [Keys]
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Key pair details
...